# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Tu Vacuna PAI — API
  description: >
    Contrato HTTP de la API de Tu Vacuna PAI. Este documento es la AUTORIDAD del
    transporte HTTP (endpoints, status codes, JSON). La semántica de la
    sincronización (comandos, estados, idempotencia, cursor, conflictos) NO se
    redefine aquí: vive en `docs/synchronization/sync-contract.md`, al que este
    documento referencia.
  version: v1
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /institutions/{institutionId}/vaccines:
    get:
      tags: [InstitutionCatalog]
      summary: Vacunas de la institucion (relaciones)
      security:
        - bearerAuth: []
      parameters:
        - { name: institutionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Relaciones vacuna-institucion
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstitutionVaccine'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /institutions/{institutionId}/vaccines/available:
    get:
      tags: [InstitutionCatalog]
      summary: Vacunas activas aun no habilitadas en la institucion
      security:
        - bearerAuth: []
      parameters:
        - { name: institutionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Vacunas disponibles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vaccine'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /institutions/{institutionId}/vaccines/{vaccineId}/enable:
    post:
      tags: [InstitutionCatalog]
      summary: Habilitar una vacuna en la institucion
      security:
        - bearerAuth: []
      parameters:
        - { name: institutionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: vaccineId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204':
          description: Vacuna habilitada
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /institutions/{institutionId}/vaccines/{vaccineId}/disable:
    post:
      tags: [InstitutionCatalog]
      summary: Deshabilitar una vacuna en la institucion
      security:
        - bearerAuth: []
      parameters:
        - { name: institutionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: vaccineId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204':
          description: Vacuna deshabilitada
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /institutions/{institutionId}/vaccines/clone:
    post:
      tags: [InstitutionCatalog]
      summary: Clonar el catalogo global en la institucion (copy-once)
      security:
        - bearerAuth: []
      parameters:
        - { name: institutionId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloneCatalogRequest'
      responses:
        '200':
          description: Resultado del clonado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloneCatalogResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /institutions/{institutionId}/vaccines/{vaccineId}/options:
    get:
      tags: [InstitutionCatalog]
      summary: Opciones operativas de la institucion
      security:
        - bearerAuth: []
      parameters:
        - { name: institutionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: vaccineId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Opciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Option'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [InstitutionCatalog]
      summary: Crear opcion operativa local
      security:
        - bearerAuth: []
      parameters:
        - { name: institutionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: vaccineId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptionRequest'
      responses:
        '201':
          description: Opcion creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Option'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /institutions/{institutionId}/vaccines/{vaccineId}/options/{optionId}:
    put:
      tags: [InstitutionCatalog]
      summary: Actualizar opcion operativa local
      security:
        - bearerAuth: []
      parameters:
        - { name: institutionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: vaccineId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: optionId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptionRequest'
      responses:
        '200':
          description: Opcion actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Option'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [InstitutionCatalog]
      summary: Desactivar opcion operativa local
      security:
        - bearerAuth: []
      parameters:
        - { name: institutionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: vaccineId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: optionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: version, in: query, required: true, schema: { type: integer, format: int64 } }
      responses:
        '204':
          description: Opcion desactivada
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /institutions/{institutionId}/vaccines/{vaccineId}/suggested-options:
    get:
      tags: [InstitutionCatalog]
      summary: Opciones sugeridas (templates globales aun no locales)
      security:
        - bearerAuth: []
      parameters:
        - { name: institutionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: vaccineId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Opciones sugeridas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Option'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /institutions/{institutionId}/vaccines/{vaccineId}/import-suggested-options:
    post:
      tags: [InstitutionCatalog]
      summary: Importar las opciones sugeridas como locales
      security:
        - bearerAuth: []
      parameters:
        - { name: institutionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: vaccineId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Opciones importadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Option'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  # ------------------------------------------------------------------- Usuarios
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido por Supabase Auth. Spring valida sub, usuario activo, permiso y scope.
  responses:
    Unauthorized:
      description: No autenticado o token inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Autenticado pero sin permiso o scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Conflicto de versión (optimistic locking) o de identidad
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
    # ---- Catalogos
...
    Vaccine:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        code: { type: string }
        category: { type: string }
        maxDoses: { type: integer }
        minAgeMonths: { type: integer, nullable: true }
        maxAgeMonths: { type: integer, nullable: true }
        hasLaboratory: { type: boolean }
        hasLot: { type: boolean }
        hasSyringe: { type: boolean }
        hasSyringeLot: { type: boolean }
        hasDiluent: { type: boolean }
        hasDropper: { type: boolean }
        hasPneumococcalType: { type: boolean }
        hasVialCount: { type: boolean }
        hasObservation: { type: boolean }
        active: { type: boolean }
        version: { type: integer }
...
    Option:
      type: object
      properties:
        id: { type: string, format: uuid }
        vaccineId: { type: string, format: uuid }
        institutionId: { type: string, format: uuid, nullable: true }
        fieldType: { type: string }
        value: { type: string }
        displayName: { type: string }
        sortOrder: { type: integer }
        isDefault: { type: boolean }
        isActive: { type: boolean }
        sourceTemplateId: { type: string, format: uuid, nullable: true }
        version: { type: integer, format: int64 }
    OptionRequest:
      type: object
      required: [fieldType, value, displayName]
      properties:
        fieldType: { type: string }
        value: { type: string }
        displayName: { type: string }
        sortOrder: { type: integer }
        isDefault: { type: boolean }
        isActive: { type: boolean }
        version: { type: integer, format: int64 }
    InstitutionVaccine:
      type: object
      properties:
        id: { type: string, format: uuid }
        institutionId: { type: string, format: uuid }
        vaccineId: { type: string, format: uuid }
        name: { type: string }
        code: { type: string }
        category: { type: string }
        enabled: { type: boolean }
        version: { type: integer, format: int64 }
    CloneCatalogRequest:
      type: object
      properties:
        includeDefaultConfig: { type: boolean }
    CloneCatalogResponse:
      type: object
      properties:
        vaccinesEnabled: { type: integer }
        optionsCopied: { type: integer }
        vaccinesTotal: { type: integer }
...